refactor(lexical-graph): move the S3 scheme and parser into s3_chunk_store
The connection-string parser and scheme constant are S3-specific, so they
live with the S3 store rather than the generic factory module, and the
scheme constant's name now says it is an S3 URI scheme.
