+++ lexical-graph/src/graphrag_toolkit/lexical_graph/storage/chunk/s3_chunk_store.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import concurrent.futures
+from urllib.parse import urlparse, parse_qs
 import logging
 from typing import Dict, List, Optional
 
@@ -11,6 +12,37 @@ from graphrag_toolkit.lexical_graph.config import GraphRAGConfig
 from graphrag_toolkit.lexical_graph.storage.chunk.chunk_store import ChunkStore
 
 logger = logging.getLogger(__name__)
+
+# The scheme that selects this store in a chunk-store connection string.
+S3_URI_SCHEME = 's3://'
+
+
+def parse_s3_connection_string(connection_string):
+    parsed = urlparse(connection_string)
+
+    bucket_name = parsed.hostname
+
+    # `s3:///prefix` parses with no hostname and would otherwise build a store
+    # against bucket None, failing later at the first call with an error that
+    # says nothing about the connection string.
+    if not bucket_name:
+        raise ValueError(
+            f'Invalid S3 connection string, no bucket name: {connection_string}. '
+            'Expected s3://bucket/prefix.'
+        )
+
+    prefix = parsed.path[1:] if parsed.path else None
+    if prefix:
+        while prefix.endswith('/'):
+            prefix = prefix[:-1]
+    prefix = prefix if prefix else None
+
+    # parse_qs always returns a list for a present key, so take the first value.
+    kms_key_arns = parse_qs(parsed.query).get('kmsKeyArn') if parsed.query else None
+    kms_key_arn = kms_key_arns[0] if kms_key_arns else None
+
+    return (bucket_name, prefix, kms_key_arn)
+
 
 # Codes S3 and S3-compatible endpoints use for an object that isn't there.
 MISSING_KEY_CODES = ('NoSuchKey', '404')
