+++ lexical-graph/src/graphrag_toolkit/lexical_graph/storage/chunk_store_factory.py
@@ -3,45 +3,20 @@
 
 import logging
 from typing import Union, Type, Dict, Optional
-from urllib.parse import urlparse, parse_qs
 
 from graphrag_toolkit.lexical_graph.storage.chunk import ChunkStore, ChunkStoreFactoryMethod
 from graphrag_toolkit.lexical_graph.storage.chunk.in_graph_chunk_store import InGraphChunkStore
-from graphrag_toolkit.lexical_graph.storage.chunk.s3_chunk_store import S3ChunkStore
+from graphrag_toolkit.lexical_graph.storage.chunk.s3_chunk_store import (
+    S3ChunkStore,
+    S3_URI_SCHEME,
+    parse_s3_connection_string,
+)
 
 logger = logging.getLogger(__name__)
 
 ChunkStoreType = Union[str, ChunkStore]
 ChunkStoreFactoryMethodType = Union[ChunkStoreFactoryMethod, Type[ChunkStoreFactoryMethod]]
 
-S3 = 's3://'
-
-
-def parse_s3_connection_string(connection_string):
-    parsed = urlparse(connection_string)
-
-    bucket_name = parsed.hostname
-
-    # `s3:///prefix` parses with no hostname and would otherwise build a store
-    # against bucket None, failing later at the first call with an error that
-    # says nothing about the connection string.
-    if not bucket_name:
-        raise ValueError(
-            f'Invalid S3 connection string, no bucket name: {connection_string}. '
-            'Expected s3://bucket/prefix.'
-        )
-
-    prefix = parsed.path[1:] if parsed.path else None
-    if prefix:
-        while prefix.endswith('/'):
-            prefix = prefix[:-1]
-    prefix = prefix if prefix else None
-
-    # parse_qs always returns a list for a present key, so take the first value.
-    kms_key_arns = parse_qs(parsed.query).get('kmsKeyArn') if parsed.query else None
-    kms_key_arn = kms_key_arns[0] if kms_key_arns else None
-
-    return (bucket_name, prefix, kms_key_arn)
 
 
 class InGraphChunkStoreFactory(ChunkStoreFactoryMethod):
@@ -72,7 +47,7 @@ class S3ChunkStoreFactory(ChunkStoreFactoryMethod):
     graph node - is still readable.
     """
     def try_create(self, chunk_info: str, **kwargs) -> Optional[ChunkStore]:
-        if not chunk_info or not chunk_info.startswith(S3):
+        if not chunk_info or not chunk_info.startswith(S3_URI_SCHEME):
             return None
 
         (bucket_name, prefix, kms_key_arn) = parse_s3_connection_string(chunk_info)
